test : test.yml
env.js : 14, 41
jest.config.js
package-lock.json
package.json : 7~
test.yml
init.js: 27
item.integration.spec.js
sequelize.js:4~

+++ src/init.js
@@ -24,7 +24,8 @@ export class Server {
   };
 
   sequelizeSync = () => {
-    return sequelize.sync({ force: true });
+    if (process.env.NODE_ENV === 'development')
+      return sequelize.sync({ force: true });
   };
 
   runServer = async () => {
